refactor(playwright-coverage): split index into types/report/collector/lifecycle modules
Extract the package internals from a single 195-line index.ts into focused
sibling modules — types, report config helpers, the per-test collector, and
the global setup/teardown lifecycle — each with co-located tests. index.ts
keeps the public API (default factory + barrel re-exports) unchanged.
