+++ packages/playwright-coverage/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@soroush.tech/playwright-coverage",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "description": "Collect per-test V8 code coverage from Playwright e2e runs and aggregate it to lcov via monocart-coverage-reports.",
   "keywords": [
     "playwright",
